requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cryptography==46.0.4
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.6
filelock==3.20.0
fsspec==2025.12.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==3.0.0
passlib==1.7.4
psutil==7.2.2
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentencepiece==0.2.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0+cpu
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
